requirements.txt
snowflake-connector-python==3.10.0
requests==2.31.0
